add Cryptographer model for tracking a user's points
The Cryptographer model has a OneToOneField with the User. A User's accumulated points is tracked using the Cryptographer to rapidly track the total number of point that they have (instead of doing a SUM every time this information is required). The many to many relationship between a challenge and a user (now a cryptographer) has been separated to its own table (manual instead of the automatic one).